requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
Django==5.2.3
django-allauth==65.9.0
django-cors-headers==4.7.0
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
mysqlclient==2.2.7
oauthlib==3.3.1
pillow==11.2.1
pycparser==2.22
PyJWT==2.9.0
requests==2.32.4
requests-oauthlib==2.0.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0